Fix cross-compilation between windows and non-windows.
Formerly, src/lib.rs used `cfg(target_os=..)` to tell which
separator to use for `env!(OUT_DIR)` and `lib.rs`.
But `target_os` only tells you which OS you're building for,
not which OS you're building on: and Rust's `include`!
interprets paths from the point of view of the _host_ OS.

With this patch, the build script now exposes an environment
variable to tell `src/lib.rs` which separator to use.

(We can't use `std::path::MAIN_SEPARATOR` directly
in `src/lib.rs`, since `concat!` doesn't accept it as an input.)

This should fix issue #15.